aws_hackathon_final/requirements.txt
Flask
yfinance
requests
beautifulsoup4
numpy
scikit-learn
google-generativeai
streamlit
boto3
